edit director

+++ controllers/directorsController.js
@@ -7,60 +7,104 @@ const fs = require('fs');
 
 class directorsController {
     addDirector(req, res) {
-
-        let dir = './public/uploads/directors';
-        if (!fs.existsSync(dir)) {
-            fs.mkdirSync(dir);
-        }
-        let storage = multer.diskStorage({
-            destination: function(req, file, callback) {
-                callback(null, dir);
-            },
-            filename: function(req, file, callback) {
-                callback(null, file.fieldname + '-' + Date.now() + file.originalname);
-            }
-        });
-        const errorFormatter = ({ location, msg, param, value, nestedErrors }) => {
-            return `${location}[${param}]: ${msg}`;
-        };
-
-        let upload = multer({
-            storage: storage,
-            limits: { fileSize: 2 * 1024 * 1024 },
-            fileFilter: function(req, file, cb) {
-                if (path.extname(file.originalname) !== ".png" && path.extname(file.originalname) !== ".jpg" && path.extname(file.originalname) !== ".jpeg") {
-                    return cb(new Error('only png or jpg'));
-                }
-
-                cb(null, true);
-            }
-        }).single('avatar');
-
-        upload(req, res, (err) => {
-            if (err) {
-                res.json(err);
-            } else {
-                if (req.file) {
-                    req.body.avatar = '/uploads/directors/' + req.file.filename;
-                } else {
-                    req.body.avatar = '';
-                }
-
-                req.checkBody('name', 'name required').notEmpty();
-                req.checkBody('lastname', 'lastname required').notEmpty();
-                req.checkBody('url', 'url must be specified').notEmpty();
-                const result = validationResult(req).formatWith(errorFormatter);
-                if (!result.isEmpty()) {
-                    // { errors: [ "body[password]: must be at least 10 chars long" ] }
-                    return res.json({ errors: result.array() });
-                } else {
-                    directors.addDirector(req.body, (err, data) => {
-                        res.redirect('/admin/adddirector');
-                    });
-                }
-            }
+        this.uploadAvatar(req,res,(err)=>{
+          if (err) {
+              res.json(err);
+          } else {
+              if (req.file) {
+                  req.body.avatar = '/uploads/directors/' + req.file.filename;
+              } else {
+                  req.body.avatar = '';
+              }
+              const errorFormatter = ({ location, msg, param, value, nestedErrors }) => {
+                  return `${location}[${param}]: ${msg}`;
+              };
+              req.checkBody('name', 'name required').notEmpty();
+              req.checkBody('lastname', 'lastname required').notEmpty();
+              req.checkBody('url', 'url must be specified').notEmpty();
+              const result = validationResult(req).formatWith(errorFormatter);
+              if (!result.isEmpty()) {
+                  // { errors: [ "body[password]: must be at least 10 chars long" ] }
+                  return res.json({ errors: result.array() });
+              } else {
+                  directors.addDirector(req.body, (err, data) => {
+                    if(err){
+                      console.log(err);
+                    }else{
+                      res.redirect('/admin/adddirector');
+                    }
+                  });
+              }
+          }
         });
     }
+
+    editDirector(req,res){
+      this.uploadAvatar(req,res,(err)=>{
+        if (err) {
+            res.json(err);
+        } else {
+            if (req.file) {
+                req.body.avatar = '/uploads/directors/' + req.file.filename;
+            } else {
+                req.body.avatar = '';
+            }
+            const errorFormatter = ({ location, msg, param, value, nestedErrors }) => {
+                return `${location}[${param}]: ${msg}`;
+            };
+            req.checkBody('id','id must me required').notEmpty();
+            req.checkBody('name', 'name required').notEmpty();
+            req.checkBody('lastname', 'lastname required').notEmpty();
+            req.checkBody('url', 'url must be specified').notEmpty();
+            const result = validationResult(req).formatWith(errorFormatter);
+            if (!result.isEmpty()) {
+                // { errors: [ "body[password]: must be at least 10 chars long" ] }
+                return res.json({ errors: result.array() });
+            } else {
+                directors.editDirector(req.body,(err,data)=>{
+                  if(err){
+                    console.log(err);
+                  }else{
+                    res.redirect(`/admin/editdirector/${data._id}`);
+                  }
+                });
+            }
+        }
+      });
+    }
+
+    uploadAvatar(req,res,cb){
+      let dir = './public/uploads/directors';
+      if (!fs.existsSync(dir)) {
+          fs.mkdirSync(dir);
+      }
+      let storage = multer.diskStorage({
+          destination: function(req, file, callback) {
+              callback(null, dir);
+          },
+          filename: function(req, file, callback) {
+              callback(null, file.fieldname + '-' + Date.now() + file.originalname);
+          }
+      });
+
+      let upload = multer({
+          storage: storage,
+          limits: { fileSize: 2 * 1024 * 1024 },
+          fileFilter: function(req, file, cb) {
+              if (path.extname(file.originalname) !== ".png" && path.extname(file.originalname) !== ".jpg" && path.extname(file.originalname) !== ".jpeg") {
+                  return cb(new Error('only png or jpg'));
+              }
+
+              cb(null, true);
+          }
+      }).single('avatar');
+
+      upload(req, res, (err) => {
+        cb(err);
+      });
+
+    }
+
     getDirectors(cb){
         directors.getDirectors(cb);
     }
@@ -75,7 +119,11 @@ class directorsController {
           if(err){
             console.log(err);
           }else{
-            fs.unlink(`public${data.avatar}`,cb);
+            if(data.avatar.length > 0 &&  fs.existsSync(`public${data.avatar}`)){
+              fs.unlink(`public${data.avatar}`,cb);
+            }else{
+              cb(null);
+            }
           }
         });
       }else{
@@ -89,7 +137,7 @@ class directorsController {
           if(err){
             console.log(err);
           }else {
-            if(fs.existsSync(`public${director.avatar}`)){
+            if(director.avatar.length > 0 && fs.existsSync(`public${director.avatar}`)){
               fs.unlink(`public${director.avatar}`,cb);
             }else{
               cb(null);
